Hack showing each line in a tr
Instead of how it currently works which is to put all the line numbers
in one td and all the code in another td. This breaks out each line into
their own tr with the line number in one td and the line of code in
another td.

This is a proof of concept not anything usable.